+++ src/argh/assembling.py
@@ -14,10 +14,11 @@ Assembling
 Functions and classes to properly assemble your commands in a parser.
 """
 import inspect
-from argparse import ZERO_OR_MORE, ArgumentParser
+from argparse import OPTIONAL, ZERO_OR_MORE, ArgumentParser
 from collections import OrderedDict
 from dataclasses import asdict
-from typing import Any, Callable, Dict, Iterator, List, Optional, Tuple, Type, Union
+from enum import Enum
+from typing import Any, Callable, Dict, Iterator, List, Optional, Tuple
 
 from argh.completion import COMPLETION_ENABLED
 from argh.constants import (
@@ -37,27 +38,92 @@ __all__ = [
     "set_default_command",
     "add_commands",
     "add_subcommands",
+    "NameMappingPolicy",
 ]
 
 
-def extract_parser_add_argument_kw_from_signature(
+class NameMappingPolicy(Enum):
+    """
+    Represents possible approaches to treat default values when inferring argument
+    specification from function signature.
+
+    * `BY_NAME_IF_KWONLY` is the default and recommended approach introduced
+      in v0.30.  It enables fine control over two aspects:
+
+      * positional vs named;
+      * required vs optional.
+
+      "Normal" arguments are identified by position, "kwonly" are identified by
+      name, regardless of the presence of default values.  A positional with a
+      default value becomes optional but still positional (``nargs=OPTIONAL``).
+      A kwonly argument without a default value becomes a required named
+      argument.
+
+      Example::
+
+          def func(alpha, beta=1, *, gamma, delta=2): ...
+
+      is equivalent to::
+
+          prog alpha [beta] --gamma [--delta DELTA]
+
+      That is, `alpha` and `--gamma` are mandatory while `beta` and `--delta`
+      are optional (they have default values).
+
+    * `BY_NAME_IF_HAS_DEFAULT` is very close to the the legacy approach
+      (pre-v0.30).  If a function argument has a default value, it becomes an
+      "option" (called by name, like ``--foo``); otherwise it's treated as a
+      positional argument.
+
+      Example::
+
+          def func(alpha, beta=1, *, gamma, delta=2): ...
+
+      is equivalent to::
+
+          prog [--beta BETA] [--delta DELTA] alpha gamma
+
+      That is, `alpha` and `gamma` are mandatory and positional, while `--beta`
+      and `--delta` are optional (they have default values).  Note that it's
+      impossible to have an optional positional or a mandatory named argument.
+
+      The difference between this policy and the behaviour of Argh before
+      v0.30 is in the treatment of kwonly arguments without default values:
+      they used to become ``--foo FOO`` (required) but for the sake of
+      simplicity they are treated as positionals.  If you are already using
+      kwonly args, please consider the better suited policy `BY_NAME_IF_KWONLY`
+      instead.
+
+    It is recommended to migrate any older code to `BY_NAME_IF_KWONLY`.
+
+    .. versionadded:: 0.30
+    """
+
+    BY_NAME_IF_HAS_DEFAULT = "specify CLI argument by name if it has a default value"
+    BY_NAME_IF_KWONLY = "specify CLI argument by name if it comes from kwonly"
+
+
+def infer_argspecs_from_function(
     function: Callable,
+    name_mapping_policy: Optional[
+        NameMappingPolicy
+    ] = NameMappingPolicy.BY_NAME_IF_KWONLY,
 ) -> Iterator[ParserAddArgumentSpec]:
     if getattr(function, ATTR_EXPECTS_NAMESPACE_OBJECT, False):
         return
 
+    if name_mapping_policy not in NameMappingPolicy:
+        raise NotImplementedError(f"Unknown name mapping policy {name_mapping_policy}")
+
     func_spec = get_arg_spec(function)
 
-    defaults: Dict[str, Any] = dict(
+    default_by_arg_name: Dict[str, Any] = dict(
         zip(reversed(func_spec.args), reversed(func_spec.defaults or tuple()))
     )
-    defaults.update(getattr(func_spec, "kwonlydefaults", None) or {})
-
-    kwonly = getattr(func_spec, "kwonlyargs", [])
 
     # define the list of conflicting option strings
     # (short forms, i.e. single-character ones)
-    named_args = set(list(defaults) + kwonly)
+    named_args = set(list(default_by_arg_name) + func_spec.kwonlyargs)
     named_arg_chars = [a[0] for a in named_args]
     named_arg_char_counts = dict(
         (char, named_arg_chars.count(char)) for char in set(named_arg_chars)
@@ -66,46 +132,69 @@ def extract_parser_add_argument_kw_from_signature(
         char for char in named_arg_char_counts if 1 < named_arg_char_counts[char]
     )
 
-    for name in func_spec.args + kwonly:
-        cli_arg_names: List[str] = []
-        akwargs: Dict[str, Any] = {}  # keyword arguments for add_argument()
-        is_required: Union[bool, Type[NotDefined]] = NotDefined
-        default_value: Any = NotDefined
+    def _make_cli_arg_names_options(arg_name) -> Tuple[List[str], List[str]]:
+        cliified_arg_name = arg_name.replace("_", "-")
+        positionals = [cliified_arg_name]
+        can_have_short_opt = arg_name[0] not in conflicting_opts
 
-        if name in defaults or name in kwonly:
-            if name in defaults:
-                default_value = defaults.get(name)
-            else:
-                is_required = True
-            cli_arg_names = [f"-{name[0]}", f"--{name}"]
-            if name.startswith(conflicting_opts):
-                # remove short name
-                cli_arg_names = cli_arg_names[1:]
-
+        if can_have_short_opt:
+            options = [f"-{cliified_arg_name[0]}", f"--{cliified_arg_name}"]
         else:
-            # positional argument
-            cli_arg_names = [name]
+            options = [f"--{cliified_arg_name}"]
 
-        # cmd(foo_bar)  ->  add_argument("foo-bar")
-        cli_arg_names = [
-            x.replace("_", "-") if x.startswith("-") else x for x in cli_arg_names
-        ]
+        return positionals, options
 
-        spec = ParserAddArgumentSpec(
-            name,
-            cli_arg_names,
-            default_value=default_value,
-            other_add_parser_kwargs=akwargs,
+    default_value: Any
+    for arg_name in func_spec.args:
+        cli_arg_names_positional, cli_arg_names_options = _make_cli_arg_names_options(
+            arg_name
         )
-        if is_required != NotDefined:
-            spec.is_required = is_required
-        yield spec
+        default_value = default_by_arg_name.get(arg_name, NotDefined)
+
+        arg_spec = ParserAddArgumentSpec(
+            func_arg_name=arg_name,
+            cli_arg_names=cli_arg_names_positional,
+            default_value=default_value,
+        )
+
+        if default_value != NotDefined:
+            if name_mapping_policy == NameMappingPolicy.BY_NAME_IF_KWONLY:
+                arg_spec.nargs = OPTIONAL
+            else:
+                arg_spec.cli_arg_names = cli_arg_names_options
+
+        yield arg_spec
+
+    for arg_name in func_spec.kwonlyargs:
+        cli_arg_names_positional, cli_arg_names_options = _make_cli_arg_names_options(
+            arg_name
+        )
+
+        if func_spec.kwonlydefaults and arg_name in func_spec.kwonlydefaults:
+            default_value = func_spec.kwonlydefaults[arg_name]
+        else:
+            default_value = NotDefined
+
+        arg_spec = ParserAddArgumentSpec(
+            func_arg_name=arg_name,
+            cli_arg_names=cli_arg_names_positional,
+            default_value=default_value,
+        )
+
+        if name_mapping_policy == NameMappingPolicy.BY_NAME_IF_KWONLY:
+            arg_spec.cli_arg_names = cli_arg_names_options
+            if default_value == NotDefined:
+                arg_spec.is_required = True
+        else:
+            if default_value != NotDefined:
+                arg_spec.cli_arg_names = cli_arg_names_options
+
+        yield arg_spec
 
     if func_spec.varargs:
-        # *args
         yield ParserAddArgumentSpec(
-            func_spec.varargs,
-            [func_spec.varargs],
+            func_arg_name=func_spec.varargs,
+            cli_arg_names=[func_spec.varargs.replace("_", "-")],
             nargs=ZERO_OR_MORE,
         )
 
@@ -129,8 +218,10 @@ def guess_extra_parser_add_argument_spec_kwargs(
       (if action was not explicitly defined).
 
     """
+    # TODO: use typing to extract
     other_add_parser_kwargs = parser_add_argument_spec.other_add_parser_kwargs
     guessed: Dict[str, Any] = {}
+    is_positional = not parser_add_argument_spec.cli_arg_names[0].startswith("-")
 
     # Parser actions that accept argument 'type'
     TYPE_AWARE_ACTIONS = "store", "append"
@@ -139,8 +230,9 @@ def guess_extra_parser_add_argument_spec_kwargs(
     default_value = parser_add_argument_spec.default_value
     if default_value not in (None, NotDefined):
         if isinstance(default_value, bool):
-            if other_add_parser_kwargs.get("action") is None:
+            if not is_positional and other_add_parser_kwargs.get("action") is None:
                 # infer action from default value
+                # (not applicable to positionals: _StoreAction doesn't accept `nargs`)
                 guessed["action"] = "store_false" if default_value else "store_true"
         elif other_add_parser_kwargs.get("type") is None:
             # infer type from default value
@@ -157,12 +249,27 @@ def guess_extra_parser_add_argument_spec_kwargs(
     return guessed
 
 
-def set_default_command(parser, function: Callable) -> None:
+def set_default_command(
+    parser,
+    function: Callable,
+    name_mapping_policy: NameMappingPolicy = NameMappingPolicy.BY_NAME_IF_KWONLY,
+) -> None:
     """
     Sets default command (i.e. a function) for given parser.
 
     If `parser.description` is empty and the function has a docstring,
     it is used as the description.
+
+    :param function:
+
+        The function to use as the command.
+
+    :name_mapping_policy:
+
+        The policy to use when mapping function arguments onto CLI arguments.
+        See :class:`.NameMappingPolicy`.
+
+        .. versionadded:: 0.30
 
     .. note::
 
@@ -183,11 +290,10 @@ def set_default_command(parser, function: Callable) -> None:
 
     declared_args: List[ParserAddArgumentSpec] = getattr(function, ATTR_ARGS, [])
     inferred_args: List[ParserAddArgumentSpec] = list(
-        extract_parser_add_argument_kw_from_signature(function)
+        infer_argspecs_from_function(function, name_mapping_policy=name_mapping_policy)
     )
 
     if declared_args and not inferred_args and not has_varkw:
-        # XXX breaking change, new behaviour
         raise AssemblingError(
             f"{function.__name__}: cannot extend argument declarations "
             "for an endpoint function that takes no arguments."
@@ -278,7 +384,7 @@ def _merge_inferred_and_declared_args(
     #     "foo-bar"     → "foo_bar"
     #
     # * argument declaration is a dictionary representing an argument;
-    #   it is obtained either from extract_parser_add_argument_kw_from_signature()
+    #   it is obtained either from infer_argspecs_from_function()
     #   or from an @arg decorator (as is).
     #
     specs_by_func_arg_name = OrderedDict()
@@ -354,6 +460,7 @@ def _is_positional(args: List[str], prefix_chars: str = "-") -> bool:
 def add_commands(
     parser: ArgumentParser,
     functions: List[Callable],
+    name_mapping_policy: NameMappingPolicy = NameMappingPolicy.BY_NAME_IF_KWONLY,
     group_name: Optional[str] = None,
     group_kwargs: Optional[Dict[str, Any]] = None,
     func_kwargs: Optional[Dict[str, Any]] = None,
@@ -374,6 +481,12 @@ def add_commands(
         concerning function signatures. The command name is inferred from the
         function name. Note that the underscores in the name are replaced with
         hyphens, i.e. function name "foo_bar" becomes command name "foo-bar".
+
+    :param name_mapping_policy:
+
+        See :class:`argh.assembling.NameMappingPolicy`.
+
+        .. versionadded:: 0.30
 
     :param group_name:
 
@@ -440,7 +553,9 @@ def add_commands(
 
         # create and set up the parser for this command
         command_parser = subparsers_action.add_parser(cmd_name, **func_parser_kwargs)
-        set_default_command(command_parser, func)
+        set_default_command(
+            command_parser, func, name_mapping_policy=name_mapping_policy
+        )
 
 
 def _extract_command_meta_from_func(func: Callable) -> Tuple[str, dict]: